Implement deletion node for splay tree and simplify toc updates
With these changes, phpinspect-buffer-parse traverses the old tree once after
each incremental parse to mark all left-over tokens as deleted. This was not
possible before, as tokens could not be deleted from the splay trees they were
in.

For toc's, it is now no longer necessary to find a common root parent to confirm
that a token has been adopted. Checking the `phpinspect-meta-deleted' slot is
enough.